revamp(Location): Move collective address rendering to own component
